Structural recovery FUCKING WORKS.
HUUUUUGE fixes since the last commit. The entire deep copy was flawed
and had to be rewritten. Also fixed a lot of errors in the main
structural recovery routine.

Now, only sequence block are implemented, but the main algorithm is
there and it is working, it is now just a matter of plug and play with
the new analyses for if blocks and loops