Things. Seriously, read full description please.
Ok this commit is strange. It started as a plain syntax fix for the
previous one (syntax fix on a 1 line comment because there was no space
after the //, so a 1 second fix. That syntax check failing the pipeline
was probably too extreme)

Then I discovered that the vector of marked nodes was clear()ed instead
of zeroed out, given that previously a set was being used, and now the
vector requires all the cells to be zeroed as a `clean` state (I dont'
exactly recall if clear() and std::fill(..,..,0) behave the same on a
vector, forgive me, I'm really tired in this moment).

At this point weird things started to happen: first SEGFAULT while
cleaning a failed CFS reconstruction. Ok, commented out those lines,
probably it is a buggy implementation (need to check tomorrow, btw).
So I commented out those line because, let's face it, that's the best
fix available at this hour of the night. And then everything exploded.

Reconstruction is failing for some reason, tried to debug with valgrind
but I realized that I'm developing at home on my macOS and valgrind is
not available at this moment on Mojave, so I spent the last 10 minutes
figuring out how to install valgrind on Mojave, and apparently the only
way is to compile from source which is taking hours on this shitty CPU.

Long story short: it is 4:20AM. I'm tired and will do the memory check
tomorrow. Goodnight.