Natural loop: break only
However, a new problem emerged: a type of do-while loop that I did not
consider and that requires additional logic to be implemented. Need to
pay attention also to the do-while hashing with the new changes
implemented. Given how natural loops are resolved, A loop should not
consider exit-points

P.S. obviously this commit won't pass