refactor: porting to rust
Currently this software is garbage. Functionalities are there, but
everything is glued together by miracle. Had to forcibly insert memory
leaks otherwise it was going to segfault everywhere. Tried to fix the
mess but it is quite complex at this point (that's why you should not
prototype in C).

Starting brand new in Rust now, so I don't have to face the same
segfault nightmare.