fix: children method and changed signatures
There was a dual indexing inside the CFG: an ID assigned to each basic
block and a index positioning the basic block in the storage array.

These two indices were used interchangeably in the exposed methods but
they really meant two different indices. Moreover for the children()
method it was stated in the doc that one type of id was returned, but
instead it was the other one.

With this commit no index is exposed at all now, and everything is done
through references. This should reduce misusage and confusion.
The BasicBlock ID itself is still present in the struct as it is needed
by the AbstractBlock class, but it will be removed in the next commit.