refactor: cfg and graph with no IDs
IDs and indices removed from Graph trait (and CFG and DirectedGraph as a
consequence). These were source of confusion and required to be careful
in order to avoid dangling indices. Everything has been rewritten now to
use Rc and storing a vector of Rc<Item> in places of the adjacency list.

Very huge refactor!