feat: add method to read CFG from file
Added a method that enables parsing a CFG previously saved by this
library/application. This is a hard requirement as additional metadata
not usually present in a .dot file is required. For this reason, the
method CFG::to_dot has been rewritten to include this metadata and the
method CFG::from_dot (and its sibling CFG::from_file) have been added.

These methods are useful to repeat analyses without having to
disassemble and extract the CFG each time.