fix: nondeterminism in nat loop resolution
Nat loop resolution was done iterating an hashmap and keeping values
with the highest number of predecessor. This was wrong in too many
levels.

Now the iteration is deterministic based on a preorder visits. (every
loop has exactly one entry point otherwise it is not solvable anyway).
Also the comparisons are deterministic and uses the offset of the BB as
a last resort.