perf: removed slow parts of the remap_nodes
The remap nodes function needed to know the old nodes to remove after a
reduction. However, getting the child of the new node was not
sufficient, in case, for example, a structure was extended. This
required a lot of iteration and preorder visits to remove and correctly
remap nodes, effectively slowing down the remap nodes to the point that
it was the slowest part of the build_cfs function (21% of its time).

Data flow have been reorganised, so the "old" nodes are now passed
around and not recomputed every time.