fix: CFG where the root is not the lowest offset
In some cases, a disassembled function has its entry point with an
offset that it is not the minimum of all offsets appearing in the CFG. I
never thought about this case and used a lot of DFS/BFS that are,
effectively, broken in this cases. With this commit, when building a
CFG, the root is not the node with the lowest offset, but the node with
the lowest amount of predecessors.