revert: root estimation only in statement CFGs
In #6329296 a new method for estimating the root CFG was added, based on
the amount of predecessors. However, in case of BareCFG I'm getting the
data directly from the disassembler, so it is actually better to ask the
disassembler directly which is the function entry point, instead of
estimating it.