fix: loops with multiple entry points
Due to an error, some loop containing multiple entry point were
correctly reduced as While or Do-While. Added a check on the
predecessors of the tail. No tests added, because soon also improper
intervals will be added (in fact the bug emerged because loop had the
priority over improper intervals).