fix: corner case in nat loop resolution
In nat loop resolution when choosing between two targets the one with
the highest depth is selected. If two targets have the same depth, the
closest to the head is selected. I assumed that this would suffice,
given that bb offsets are unique, however, I didn't consider the case
where the head can be exactly in-between the two targets, at the same
distance and the both have the same depth (this seems very unlikely but
unfortunately it is not).